Implementation of being able to spend coins in order not to lose when a prosecution fails.
The player elimination system had to be changed by removing it from its own function of checking whether an accusation is true or not. Now when you fail you don't lose directly.

The use of the previous function has been modified in the tests, in the text section and in the graphic section of the application.